fix multiple minor issues (c-style casts, default branch, etc.)

+++ src/model/SubaddressAccountModel.cpp
@@ -53,7 +53,7 @@ int SubaddressAccountModel::rowCount(const QModelIndex &) const
 
 QVariant SubaddressAccountModel::data(const QModelIndex &index, int role) const
 {
-    if (!index.isValid() || index.row() < 0 || (unsigned)index.row() >= m_subaddressAccount->count())
+    if (!index.isValid() || index.row() < 0 || static_cast<quint64>(index.row()) >= m_subaddressAccount->count())
         return {};
 
     QVariant result;
